+++ imbot/auth_test.go
