+++ imbot/imbot.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/go-telegram/bot/models"
 	"github.com/tingly-dev/tingly-box/imbot/core"
+	"github.com/tingly-dev/tingly-box/imbot/platform/feishu"
 	"github.com/tingly-dev/tingly-box/imbot/platform/telegram"
 )
 
@@ -447,3 +448,28 @@ func GetPlatformBehavior(platform Platform) core.PlatformBehavior {
 
 // PlatformBehavior re-exports the platform behavior record.
 type PlatformBehavior = core.PlatformBehavior
+
+// commandMenuSetup maps a platform to its native command-menu installer.
+//
+// This lives beside the bot-creator registry because it is the same kind of
+// fact: what a platform can do and how to reach it. It used to be a switch in
+// internal/remote_control's bot manager, which meant the consumer had to
+// import platform packages and be edited whenever a platform gained a menu.
+//
+// A platform with no entry simply has no native command menu — that is a
+// normal outcome, not an error.
+var commandMenuSetup = map[core.Platform]func(core.Bot, *core.CommandRegistry) error{
+	core.PlatformTelegram: telegram.SetupMenuButton,
+	core.PlatformFeishu:   feishu.SetupQuickActions,
+	core.PlatformLark:     feishu.SetupQuickActions,
+}
+
+// SetupCommandMenu installs the command registry into the platform's native
+// menu. Platforms without one are a no-op, so callers do not need to ask first.
+func SetupCommandMenu(bot core.Bot, platform core.Platform, reg *core.CommandRegistry) error {
+	setup, ok := commandMenuSetup[platform]
+	if !ok {
+		return nil
+	}
+	return setup(bot, reg)
+}
